Fixed WinRT build

+++ cpp/src/IceWS/Util.cpp
@@ -31,14 +31,16 @@ IceWS::calcSHA1(const vector<unsigned char>& data)
     ::SHA1(&data[0], static_cast<unsigned long>(data.size()), &hash[0]);
     return hash;
 #else
-    auto dataA = ref new Platform::Array<unsigned char>(const_cast<unsigned char*>(&data[0]), data.size());
+    auto dataA = ref new Platform::Array<unsigned char>(const_cast<unsigned char*>(&data[0]),
+						        static_cast<unsigned int>(data.size()));
     auto hasher = Windows::Security::Cryptography::Core::HashAlgorithmProvider::OpenAlgorithm("SHA1");
     auto hashed = hasher->HashData(Windows::Security::Cryptography::CryptographicBuffer::CreateFromByteArray(dataA));
     auto reader = ::Windows::Storage::Streams::DataReader::FromBuffer(hashed);
     std::vector<unsigned char> result(reader->UnconsumedBufferLength);
     if(!result.empty())
     {
-        reader->ReadBytes(::Platform::ArrayReference<unsigned char>(&result[0], result.size()));
+        reader->ReadBytes(::Platform::ArrayReference<unsigned char>(&result[0], 
+								    static_cast<unsigned int>(result.size())));
     }
     return result;
 #endif
